requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
Flask-Dance==6.2.0          # For Google OAuth
Authlib==1.3.0              # Optional: more flexible OAuth (can use one or both)
gunicorn==21.2.0            # Production WSGI server for Render
psycopg2-binary==2.9.10     # PostgreSQL driver for SQLAlchemy
python-dotenv==1.0.1        # Loads local .env for dev (SQLite or local DB)
requests==2.31.0            # For syncing data from local app
Werkzeug==2.3.8
